FEAT: GameSettingRequest validation
gameSetting 수정 메소드에 validation 적용

+++ backend/arc205/src/main/java/com/project/arc205/game/gamedata/model/entity/GameSetting.java
@@ -10,8 +10,8 @@ public class GameSetting {
     private Integer maxPlayers = 6;
     private Integer numberOfMafias = 1;
     private Integer numberOfMissions = 1;
-    private Integer visionRange = 1;
-    private Integer playerSpeed = 1;
+    private Double visionRange = 1.0;
+    private Double playerSpeed = 1.0;
     private Integer totalConferenceTime = 20;
     private Integer voteProgressTime = 10;
